Change some pointer alignment checks in irregexp to use intptr_t.
Review URL: http://codereview.chromium.org/109022


git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@1866 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/interpreter-irregexp.cc
@@ -130,13 +130,13 @@ static void TraceInterpreter(const byte* code_base,
 
 
 static int32_t Load32Aligned(const byte* pc) {
-  ASSERT((reinterpret_cast<int>(pc) & 3) == 0);
+  ASSERT((reinterpret_cast<intptr_t>(pc) & 3) == 0);
   return *reinterpret_cast<const int32_t *>(pc);
 }
 
 
 static int32_t Load16Aligned(const byte* pc) {
-  ASSERT((reinterpret_cast<int>(pc) & 1) == 0);
+  ASSERT((reinterpret_cast<intptr_t>(pc) & 1) == 0);
   return *reinterpret_cast<const uint16_t *>(pc);
 }
